requirements.txt
alembic==1.5.8
bcrypt==3.2.0
beautifulsoup4==4.9.3
boto3==1.17.40
botocore==1.20.41
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.14.4
chardet==4.0.0
click==8.0.4
cryptography==41.0.3
dataclasses==0.6
Flask==1.1.2
Flask-JWT-Extended==3.25.0
Flask-Migrate==4.0.1
Flask-SQLAlchemy==2.5.1
idna==2.10
importlib-metadata==6.0.0
itsdangerous==2.0.1
Jinja2==2.11.3
jmespath==0.10.0
Mako==1.2.3
MarkupSafe==2.0.1
numpy==1.20.2
pandas==1.4.4
pip==21.0.1
pyarrow==3.0.0
pycparser==2.21
PyJWT==1.7.1
pymongo==3.11.3
PyMySQL==1.0.2
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2023.3
requests==2.25.1
s3transfer==0.3.7
setuptools
six==1.16.0
soupsieve==2.4
SQLAlchemy==1.3.23
urllib3==1.26.4
waitress==2.0.0
Werkzeug==2.0.3
wheel==0.36.2
zipp==3.11.0
redis==5.0.0
async-timeout==4.0.3
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0